\name{Rcpuid-package}
\alias{Rcpuid-package}
\docType{package}
\title{Interface to the cpuid library}
\description{
  This package offers several utilities for looking up
  basic CPU information, such as the number of cores, cpu clock,
  and an estimated peak number of FLOPS.
}
\details{
\tabular{ll}{
  Package: \tab Rcpuid\cr
  Type: \tab Package\cr
  License: \tab BSD 2-Clause\cr
}
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
