% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cpuid.r
\name{cpu_clock}
\alias{cpu_clock}
\title{CPU Clock}
\usage{
cpu_clock()
}
\value{
Returns a list containing:
\tabular{ll}{
   ncores \tab Number of "cores" (physical + logical). \cr
   clock.os \tab The CPU clock as reported by the OS. \cr
   clock.tested \tab The CPU clock determined from a small test. \cr
   peak \tab Theoretical peak floating point operations per second. \cr
}
}
\description{
.
}
\details{
.
}
\examples{
\dontrun{
library(Rcpuid, quietly=TRUE)
cpu_clock()
}
}
\seealso{
\code{\link{cpu_id}, \link{cpu_ins}}
}

