% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cpuid.r
\name{cpu_id}
\alias{cpu_id}
\title{CPU ID}
\usage{
cpu_id()
}
\value{
Returns a list containing:
\tabular{ll}{
    vendor \tab CPU vendor \cr
    codename \tab Vendor's CPU codename \cr
    brand \tab CPU The CPU brand label \cr
}
}
\description{
A utility to look up some basic information about the CPU.
}
\examples{
\dontrun{
library(Rcpuid, quietly=TRUE)
cpu_id()
}
}
\seealso{
\code{\link{cpu_ins}, \link{cpu_clock}}
}

