\name{flops}
\alias{flops}

\title{ Flops Constructor }

\description{
  .
}
\usage{
  flops(size=0, unit="best")
}
\arguments{
  \item{size}{.}
  \item{unit}{.}
}
\details{
  .
}
\value{
  Returns a flops object.
}
\seealso{
  \code{\link{cpu_clock}}
}
\examples{
\dontrun{
library(Rcpuid, quietly=TRUE)
flops(2000000) # 2 MFLOPS
}
}


