% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linpack.r
\name{linpack}
\alias{linpack}
\alias{linpack_benchmark}
\title{Linpack}
\usage{
linpack(A, B, solvefun)

linpack_benchmark()
}
\arguments{
\item{A}{The square matrix representing the system of equations.}

\item{B}{The "right hand side", of the same order as A.}
}
\description{
.

TODO
}
\details{
.
}
\examples{
\dontrun{
library(Rcpuid, quietly=TRUE)
flops(2000000) # 2 MFLOPS
}
}
\seealso{
\code{\link{solve_nocopy}}
}

