% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flops.r
\name{print-flops}
\alias{print-flops}
\alias{print.flops}
\title{print-flops}
\usage{
\method{print}{flops}(x, ..., unit = "best", digits = 3)
}
\arguments{
\item{x}{A flops object.}

\item{...}{Ignored.}

\item{unit}{The flops display unit to use.}

\item{digits}{The number of decimal digits to display.}
}
\description{
print-flops
}

