% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linpack.r
\name{print.linpack}
\alias{print.linpack}
\title{Prints linpack objects}
\usage{
print.linpack(x, ..., digits = 3)
}
\arguments{
\item{x}{A linpack object to print.}

\item{...}{Extra arguments (ignored).}

\item{digits}{Number of digits to print.}
}
\description{
Prints linpack objects
}

