% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linpack.r
\name{solve_nocopy}
\alias{solve_nocopy}
\title{No-Copy Solve}
\usage{
solve_nocopy(A, B)
}
\arguments{
\item{A}{The square matrix representing the system of equations.}

\item{B}{The "right hand side", of the same order as A.}
}
\description{
Like R's \code{solve()}, but without copying the data matrix \code{A}.
}
\details{
WARNING: only use for benchmarking, or if you really know what you're doing.
Calling then and then referencing \code{A} can be dangerous.

This is a non-distributed "solvefun" for \code{linpack()}.
}
\seealso{
\code{\link{linpack}}
}

