% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linpack.r
\name{[.linpack}
\alias{[.linpack}
\title{Subsets linpack outputs.}
\usage{
\method{[}{linpack}(x, i)
}
\arguments{
\item{x}{linpack object.}

\item{i}{Index.

rdname subset}
}
\description{
Subsets linpack outputs.
}

