% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpack.r
\name{subset-linpack}
\alias{[.linpack}
\alias{subset-linpack}
\title{Subsets linpack outputs.}
\usage{
\method{[}{linpack}(x, i)
}
\arguments{
\item{x}{linpack object.}

\item{i}{Index.}
}
\description{
Subsets linpack outputs.
}

