% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flops.r
\name{unflop}
\alias{unflop}
\title{unflop}
\usage{
unflop(x)
}
\arguments{
\item{x}{A flops object.}
}
\value{
The number of flops represented by x.
}
\description{
Convert a flops object into a numeric.
}
\details{
This function differs from a simple \code{as.numeric()} call in
that flops objects store the unit flops.  So if you have a flops
object \code{x} that prints "10 MFLOPS", \code{as.numeric(x)}
would produce 10, while \code{unflop()} would produce 10000000.
}

