% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cpuid.r
\name{cpu_clock}
\alias{cpu_clock}
\title{CPU Clock}
\usage{
cpu_clock(milis = 200, quad_check = FALSE)
}
\arguments{
\item{milis}{The amount of time in miliseconds the clock test should run for.
Recommended range is 10-1000.}

\item{quad_check}{logical; if \code{TRUE}, the average of a best of four runs is
returned.}
}
\value{
Returns a list containing:
\tabular{ll}{
   ncores \tab Number of "cores" (physical + logical). \cr
   clock.os \tab The CPU clock as reported by the OS. \cr
   clock.tested \tab The CPU clock determined from a small test. \cr
   peak \tab Theoretical peak floating point operations per second. \cr
}
}
\description{
A utility to look up some basic performance information about
the CPU.
}
\details{
There may be a large difference between the CPU clock as reported
by the OS and as discovered by the small test.  Modern operating
systems keep the CPU clock scaled down to save power, and scale
them up as needed.
}
\examples{
\dontrun{
library(okcpuid, quietly=TRUE)
cpu_clock()
}
}
\seealso{
\code{\link{cpu_id}, \link{cpu_ins}}
}

