% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flops.r
\name{flops}
\alias{flops}
\title{Flops Constructor}
\usage{
flops(size = 0, unit = "best")
}
\arguments{
\item{size}{}

\item{unit}{}
}
\value{
Returns a flops object.
}
\description{
.
}
\details{
.
}
\examples{
\dontrun{
library(Rcpuid, quietly=TRUE)
flops(2000000) # 2 MFLOPS
}
}

